+++ noodles-csi/src/async/io.rs
@@ -1,7 +1,8 @@
 //! Async CSI.
 
+mod indexed_reader;
 mod query;
 mod reader;
 mod writer;
 
-pub use self::{query::Query, reader::Reader, writer::Writer};
+pub use self::{indexed_reader::IndexedReader, query::Query, reader::Reader, writer::Writer};